+++ test/Editing.test.js
@@ -1,18 +1,6 @@
 /* eslint-disable no-use-before-define */
 import { module } from 'substance-test'
-
-import Document from '../model/Document'
-import DocumentSchema from '../model/DocumentSchema'
-import EditorSession from '../model/EditorSession'
-import BlockNode from '../model/BlockNode'
-import InlineNode from '../model/InlineNode'
-import AbstractEditor from '../ui/AbstractEditor'
-import Component from '../ui/Component'
-import ContainerEditor from '../ui/ContainerEditor'
-import Configurator from '../util/Configurator'
-import ParagraphPackage from '../packages/paragraph/ParagraphPackage'
-import StrongPackage from '../packages/strong/StrongPackage'
-import ListPackage from '../packages/list/ListPackage'
+import setupEditor from './fixture/setupEditor'
 
 const test = module('Editing')
 
@@ -995,69 +983,6 @@ test.UI("[L6-3]: Merging a List into previous List using DELETE", function(t) {
 //   t.end()
 // })
 
-class TestEditor extends AbstractEditor {
-
-  constructor(...args) {
-    super(...args)
-    this.handleActions({
-      domSelectionRendered: function() {}
-    })
-  }
-
-  render($$) {
-    let doc = this.editorSession.getDocument()
-    let el = $$('div')
-    let body = $$(ContainerEditor, {
-      node: doc.get('body')
-    })
-    el.append(body)
-    return el
-  }
-}
-
-function setupEditor(t, ...f) {
-  let editor = TestEditor.mount({ editorSession: fixture(...f) }, t.sandbox)
-  let editorSession = editor.editorSession
-  let doc = editorSession.getDocument()
-  return { editor, editorSession, doc }
-}
-
-function getConfig() {
-  let config = new Configurator()
-  config.addToolGroup('annotations')
-  config.defineSchema(new DocumentSchema('test-article', 1.0, {
-    defaultTextType: 'paragraph'
-  }))
-  config.import(ParagraphPackage)
-  config.import(StrongPackage)
-  config.import(ListPackage)
-  config.addNode(TestBlockNode)
-  config.addNode(TestInlineNode)
-  config.addComponent('test-block', Component)
-  return config
-}
-
-function fixture(...args) {
-  let config = getConfig()
-  let doc = new Document(config.getSchema())
-  let body = doc.create({
-    type: 'container',
-    id: 'body'
-  })
-  args.forEach((seed)=>{
-    seed(doc, body)
-  })
-  let editorSession = new EditorSession(doc, { configurator: config })
-  editorSession.setSelection({
-    type: 'property',
-    path: ['p1', 'content'],
-    startOffset: 0,
-    containerId: 'body',
-    surfaceId: 'body'
-  })
-  return editorSession
-}
-
 const P1_TEXT = 'p1:abcdef'
 
 function _p1(doc, body) {
@@ -1174,12 +1099,3 @@ function _block2(doc, body) {
   })
   body.show('block2')
 }
-
-class TestInlineNode extends InlineNode {}
-TestInlineNode.type = 'test-inline'
-TestInlineNode.schema = {
-  foo: { type: 'string' }
-}
-
-class TestBlockNode extends BlockNode {}
-TestBlockNode.type = 'test-block'
